BC-6449 - switch school update to api v3
